Refactor authorization challenges to its own package
Split challenges into its own package. Avoids possible
import cycle with challenges from client.
