+++ daemon/internal/registryclient/auth/challenge/authchallenge_test.go
@@ -1,4 +1,4 @@
-package auth
+package challenge
 
 import (
 	"fmt"
@@ -50,7 +50,7 @@ func TestAuthChallengeNormalization(t *testing.T) {
 
 func testAuthChallengeNormalization(t *testing.T, host string) {
 
-	scm := NewSimpleChallengeManager()
+	scm := NewSimpleManager()
 
 	url, err := url.Parse(fmt.Sprintf("http://%s/v2/", host))
 	if err != nil {
@@ -86,7 +86,7 @@ func testAuthChallengeNormalization(t *testing.T, host string) {
 
 func testAuthChallengeConcurrent(t *testing.T, host string) {
 
-	scm := NewSimpleChallengeManager()
+	scm := NewSimpleManager()
 
 	url, err := url.Parse(fmt.Sprintf("http://%s/v2/", host))
 	if err != nil {
